Add routing and wire up sorting
This introduces react-router so that we can keep the sorting state in the address bar.

At this point it made sense to extract everything specific to the table into a component to manage the rendering of the entire page. This makes app.js all about bootstraping our providers and setting up some routes as it should be